Bare Bones Prototype
After playing around with Spritelib, I have come up with prototype,
which is essentially as barebones as you can get with the program.

I've begun to create the character class using a sprite off the
internet, this is where I'll further try and get the 'moving character'.
Sorry I'm not as good at programming as other people, working on
catching up to the current level.

Also, a couple of other goodies are attached.